+++ clients/privacy-center/types/api/models/Page_ClassifyInstanceResponseValues_.ts
@@ -1,0 +1,13 @@
+/* istanbul ignore file */
+/* tslint:disable */
+/* eslint-disable */
+
+import type { ClassifyInstanceResponseValues } from "./ClassifyInstanceResponseValues";
+
+export type Page_ClassifyInstanceResponseValues_ = {
+  items: Array<ClassifyInstanceResponseValues>;
+  total: number | null;
+  page: number | null;
+  size: number | null;
+  pages?: number | null;
+};