+++ clients/privacy-center/types/api/models/SecurityApplicationConfig.ts
@@ -1,0 +1,10 @@
+/* istanbul ignore file */
+/* tslint:disable */
+/* eslint-disable */
+
+export type SecurityApplicationConfig = {
+  /**
+   * A list of client addresses allowed to communicate with the Fides webserver.
+   */
+  cors_origins?: Array<string> | null;
+};